all good so far

+++ ft_putstr_fd.c
@@ -6,17 +6,17 @@
 /*   By: mgallizz <mgallizz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/14 13:36:10 by mgallizz          #+#    #+#             */
-/*   Updated: 2020/12/14 14:17:48 by mgallizz         ###   ########.fr       */
+/*   Updated: 2020/12/14 23:14:23 by mgallizz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putstr_fd(char *s, int fd)
+void		ft_putstr_fd(char *s, int fd)
 {
-	int oct;
-	
-	if (!s || fd < 0)
+	size_t	oct;
+
+	if (!s || (fd < 0))
 		return ;
 	oct = ft_strlen(s);
 	write(fd, s, oct);
